# swagger.yaml
openapi: 3.0.0
info:
  title: Mi E-Commerce
  version: 2.0.0
  description: |-
    Esta API esta orientada a interactuar con el front de una pagina web de E-Commerce del grupo Cencosud SA.
    Git Repo: https://github.com/rodrigobenta/segundoSprint  
servers:
  - url: http://localhost:3000/api/v2
    description: Entorno Desarrollo
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /users/login/:
    post:
      tags: 
        - usuarios
      summary: Login al sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                username:
                  type: string
                password:
                  type: string
              
      responses: 
        '200':
          description: logueado con exito
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: number
                      username:
                        type: string
                  token:
                    type: string
        '404':
          description: No se encuentra el usuario
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
  /login/:
    post:
      tags: 
        - usuarios
      summary: Login al sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                username:
                  type: string
                password:
                  type: string
              
      responses: 
        '200':
          description: logueado con exito
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: number
                      username:
                        type: string
                  token:
                    type: string
        '404':
          description: No se encuentra el usuario
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string                  
  /users/:
    get:
      tags: 
        - usuarios
      summary: Lista los usuarios
      security:
        - bearerAuth: []
      
      responses: 
        '200':
          description: ok
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: object
                    properties:
                      id:
                        type: number
                      email:
                        type: string
                      username:
                        type: string
                      password:
                        type: string
                      firstname:
                        type: string
                      lastname:
                        type: string
                      profilepic:
                        type: string
                      role:
                        type: string

        
        '401':
          description: No posee los permisos para acceder a la funcionalidad
          content:
            application/json:
              schema:
                properties:
                  Mensaje:
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje:
                    type: string
    post:
      tags: 
        - usuarios
      summary: Crear un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
                firstname:
                  type: string
                lastname:
                  type: string
                profilepic:
                  type: string
                role:
                  type: string

      responses: 
        '201':
          description: Usuario creado con exito
          content:
            application/json:
              schema:
                properties:
                  userCreated:
                    type: object
                    properties:
                      id:
                        type: number
                      email:
                        type: string
                      username:
                        type: string
                      firstname:
                        type: string
                      lastname:
                        type: string
                      profilepic:
                        type: string
                      role:
                        type: string
                      cart:
                        type: array
                        items:
                          type: object
                          properties:
                            product:
                              type: number
                            quantity:
                              type: number

        '400':
          description: Username, Email o Campos faltantes
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string

        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string
...
  /products:
    get:
      tags: 
        - productos
      summary: Lista todos los productos 
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          required: false
          description: busca los productos segun su categoria, sino, por defecto, lista todos los productos
      
      responses: 
        '200':
          description: Listado de productos
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                        title:
                          type: string
                        price:
                          type: number
                        description:
                          type: string
                        image:
                          type: string
                        gallery:
                          type: array
                          items:
                            type: object 
                            properties:
                              picture_id:
                                type: number
                              picture_url:
                                type: string
                              description:
                                type: string
                        category:
                          type: string        
                        mostwanted:
                          type: boolean
                        stock:
                          type: number

        '404':
          description: Productos con dicha categoria no encontrados
          content:
            application/json:
              schema:
                properties:
                  msg:
                    type: string                      
      
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
    post:
      tags: 
        - productos
      summary: Crea un producto 
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                stock:
                  type: number
                description:
                  type: string
                fk_id_category:
                  type: string
                mostwanted:
                  type: number

      
      responses: 
        '201':
          description: Producto creado con exito
          content:
            application/json:
              schema:
                properties:
                  list:
                    type: object
                    properties:
                      id:
                        type: number
                      title:
                        type: string
                      price:
                        type: number
                      description:
                        type: string
                      image:
                        type: string
                      category:
                        type: string
                      gallery:
                        type: array
                        items:
                            type: number
                      mostwanted:
                        type: boolean
                      stock:
                        type: number

        '400':
          description: Faltan campos necesarios para crear el producto, o el producto ya existe.
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
...
  /pictures:
    post:
      tags: 
        - pictures
      summary: Crea una nueva foto para un producto 
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                url:
                  type: string
                description:
                  type: string
                fk_id_product:
                  type: number
      
      responses: 
        '201':
          description: Foto creada con exito
          content:
            application/json:
              schema:
                properties:
                      id:
                        type: number
                      url:
                        type: string
                      description:
                        type: string
                      fk_id_product:
                        type: string
  
        '400':
          description: bad request
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
    
    get:
      tags: 
        - pictures
      summary: Busca las fotos de un producto identificado con un Id 
      security:
        - bearerAuth: []
      parameters:
        - name: product
          in: query
          schema:
            type: string
          required: false
          description: Id del producto
      
      responses: 
        '200':
          description: Fotos listadas con exito
          content:
            application/json:
              schema:
                properties:
                  
                      id:
                        type: number
                      url:
                        type: string
                      description:
                        type: string

        '400':
          description:  bad request
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
        
        '500':
          description: Server error
          content:
            application/json:
              schema:
                properties:
                  Mensaje: 
                    type: string
...